Revert "Split functionality up."
This reverts commit 67bb527dbab0321797dcd57b4b9a4263bf832d53.
